+++ src/fields/bone.rs
@@ -10,7 +10,7 @@ pub struct Bone {
 }
 
 impl Bone {
-    pub const ID: u32 = MdlxMagic::BONE as u32;
+    pub const ID: u32 = MdlxMagic::BONE;
 
     pub fn read_mdx(cur: &mut Cursor<&Vec<u8>>) -> Result<Self, MyError> {
         Ok(Self { base: Node::read_mdx(cur)?, geoset_id: cur.readx()?, geoanim_id: cur.readx()? })
