+++ src/fields/eventobj.rs
@@ -7,7 +7,7 @@ pub struct EventObject {
 }
 
 impl EventObject {
-    pub const ID: u32 = MdlxMagic::EVTS as u32;
+    pub const ID: u32 = MdlxMagic::EVTS;
 
     pub fn read_mdx(cur: &mut Cursor<&Vec<u8>>) -> Result<Self, MyError> {
         let mut this = Build! { base: Node::read_mdx(cur)? };
