+++ src/fields/helper.rs
@@ -6,7 +6,7 @@ pub struct Helper {
 }
 
 impl Helper {
-    pub const ID: u32 = MdlxMagic::HELP as u32;
+    pub const ID: u32 = MdlxMagic::HELP;
 
     pub fn read_mdx(cur: &mut Cursor<&Vec<u8>>) -> Result<Self, MyError> {
         Ok(Self { base: Node::read_mdx(cur)? })
